Move FeebDB types to their own modules
They were temporarily together for ease of development, but now the API
has reached a stable state that we can probably move them into their own
modules.
